requirements.txt
streamlit
opencv-python
scikit-image
numpy
scipy
pillow
